Added Log out option to drop down
I added a log out option to the drop down menu for both staff and managers.
I did this because I knew in the demo and interview that I would need to change between accounts and making it go smoother would make it easier considering
We lost 50% of our time for the interviews (30min -> 15min).